JWT initialization with access to the API for admin only

+++ app/API/users_api.py
@@ -2,6 +2,7 @@
 
 from flask import Blueprint, jsonify, request, abort
 from app.models import User, Apartment
+from flask_jwt_extended import create_access_token, jwt_required, get_jwt_identity
 from app import db, bcrypt
 
 users_api_bp = Blueprint('users_api', __name__, url_prefix='/api')
@@ -21,27 +22,46 @@ def serialize_user(user):
 
 
 @users_api_bp.route('/users', methods=['GET'])
+@jwt_required()
 def get_users():
     users = User.query.all()
     return jsonify([serialize_user(user) for user in users]), 200
 
 
 @users_api_bp.route('/users/<int:user_id>', methods=['GET'])
+@jwt_required()
 def get_user(user_id):
     user = User.query.get_or_404(user_id)
-    print(user.check_password('password123'))
     return jsonify(serialize_user(user)), 200
 
 
+@users_api_bp.route('/login', methods=['POST'])
+def login():
+    data = request.get_json()
+    email = data.get('Email')
+    password = data.get('Password')
+
+    user = User.query.filter_by(Email=email).first()
+    if user and bcrypt.check_password_hash(user.Password, password):
+        if user.UserType != 'admin':
+            return jsonify({"msg": "Access denied: admin only"}), 403
+        access_token = create_access_token(identity={'email': user.Email})
+        return jsonify(access_token=access_token), 200
+    return jsonify({"msg": "Invalid email or password"}), 401
+
+
 @users_api_bp.route('/users', methods=['POST'])
+@jwt_required()
 def create_user():
     data = request.json
+    if User.query.filter_by(Email=data['Email']).first():
+        return jsonify({'message': 'User with this email already exists'}), 409
     new_user = User(
         Email=data['Email'],
         Password=data['Password'],
         IsStudent=data['IsStudent'],
         Name=data['Name'],
-        PhoneNumber=data['PhoneNumber'],
+        PhoneNumber=data.get('PhoneNumber', None),
         UserType=data['UserType'],
         UserImage=data.get('UserImage', None)
     )
@@ -51,9 +71,12 @@ def create_user():
 
 
 @users_api_bp.route('/users/<int:user_id>', methods=['PUT'])
+@jwt_required()
 def update_user(user_id):
     user = User.query.get_or_404(user_id)
     data = request.json
+    if get_jwt_identity() != user.UserID:
+        return jsonify({'message': 'Permission denied'}), 403
 
     user.Email = data.get('Email', user.Email)
     user.Password = data.get('Password', user.Password)
@@ -68,8 +91,11 @@ def update_user(user_id):
 
 
 @users_api_bp.route('/users/<int:user_id>', methods=['DELETE'])
+@jwt_required()
 def delete_user(user_id):
     user = User.query.get_or_404(user_id)
+    if get_jwt_identity() != user.UserID:
+        return jsonify({'message': 'Permission denied'}), 403
     db.session.delete(user)
     db.session.commit()
     return jsonify({'message': 'User deleted successfully', 'UserID': user.UserID}), 200
